+++ Resources/Prototypes/GameRules/pests.yml
@@ -27,14 +27,17 @@
       path: /Audio/Announcements/attention.ogg
     earliestStart: 15
     weight: 6
-    duration: 30 # DeltaV: was 50, used as a delay now
+    duration: 50
   - type: VentCrittersRule
-    table: !type:GroupSelector # DeltaV: EntityTable instead of spawn entries
-      children:
-      - id: MobMouse
-      - id: MobMouse1
-      - id: MobMouse2
-      #- id: MobMouseCancer # starcup: disable cancer mouse spawn
+    entries:
+    - id: MobMouse
+      prob: 0.02
+    - id: MobMouse1
+      prob: 0.02
+    - id: MobMouse2
+      prob: 0.02
+    # - id: MobMouseCancer # starcup: disable cancer mouse spawn
+    #   prob: 0.001
 # Events always spawn a critter regardless of Probability https://github.com/space-wizards/space-station-14/issues/28480 I added the Rat King to their own event with a player cap.
 
 - type: entity
@@ -47,15 +50,21 @@
       path: /Audio/Announcements/attention.ogg
     earliestStart: 15
     weight: 6
-    duration: 30 # DeltaV: was 50, used as a delay now
+    duration: 50
     minimumPlayers: 30 # Hopefully this is enough for the Rat King's potential Army (it was not, raised from 15 -> 30)
   - type: VentCrittersRule
-    table: !type:GroupSelector # DeltaV: EntityTable instead of spawn entries
-      children:
-      - id: MobMouse
-      - id: MobMouse1
-      - id: MobMouse2
-      #- id: MobMouseCancer # starcup: disable cancer mouse spawn
+    entries:
+    - id: MobMouse
+      prob: 0.02
+    - id: MobMouse1
+      prob: 0.02
+    - id: MobMouse2
+      prob: 0.02
+    - id: MobMouseCancer
+      prob: 0.001
+    specialEntries:
+    - id: SpawnPointGhostRatKing
+      prob: 0.001
     specialEntries:
     - id: SpawnPointGhostRatKing
       prob: 0.001
@@ -69,16 +78,15 @@
     startAudio:
       path: /Audio/Announcements/attention.ogg
     weight: 6
-    duration: 30 # DeltaV: was 50, used as a delay now
+    duration: 50
   - type: VentCrittersRule
-    table: !type:GroupSelector # DeltaV: EntityTable instead of spawn entries
-      children:
-      - id: MobCockroach
-        weight: 0.79
-      - id: MobMothroach
-        weight: 0.206 # starcup: 0.21 -> 0.206, DeltaV doesn't have moproaches
-      - id: MobMoproach
-        prob: 0.004
+    entries:
+    - id: MobCockroach
+      prob: 0.03
+    - id: MobMothroach
+      prob: 0.008
+    - id: MobMoproach
+      prob: 0.004
 
 - type: entity
   id: SnailMigrationLowPop
@@ -89,16 +97,15 @@
     startAudio:
       path: /Audio/Announcements/attention.ogg
     weight: 6
-    duration: 30 # DeltaV: was 50, used as a delay now
+    duration: 50
   - type: VentCrittersRule
-    table: !type:GroupSelector # DeltaV: EntityTable instead of spawn entries
-      children: # starcup: add weights
-      - id: MobSnail
-        weight: 0.92
-      - id: MobSnailSpeed
-        weight: 0.04
-      - id: MobSnailMoth
-        weight: 0.04
+    entries:
+    - id: MobSnail
+      prob: 0.02
+    - id: MobSnailSpeed
+      prob: 0.002
+    - id: MobSnailMoth
+      prob: 0.002
 
 - type: entity
   id: SnailMigration
@@ -110,16 +117,15 @@
       path: /Audio/Announcements/attention.ogg
     earliestStart: 15
     weight: 6
-    duration: 30 # DeltaV: was 50, used as a delay now
+    duration: 50
     minimumPlayers: 30
   - type: VentCrittersRule
-    table: !type:GroupSelector # DeltaV: EntityTable instead of spawn entries
-      children:
-      - id: MobSnail
-        weight: 0.84
-      - id: MobSnailSpeed
-        weight: 0.08
-      - id: MobSnailMoth
-        weight: 0.08
-      # - id: MobSnailInstantDeath # starcup: disable instant death snail
-      #   prob: 0.00001 #  ~ 1:2000 snails
+    entries:
+    - id: MobSnail
+      prob: 0.02
+    - id: MobSnailSpeed
+      prob: 0.002
+    - id: MobSnailMoth
+      prob: 0.002
+    # - id: MobSnailInstantDeath # starcup: disable instant death snail
+    #   prob: 0.00001 #  ~ 1:2000 snails
