+++ Resources/Prototypes/_starcup/Entities/Structures/Piping/Atmospherics/unary.yml
@@ -1,7 +1,7 @@
 - type: entity
   parent: GasVentPump
   id: GasVentPumpNoSpawns
-  suffix: No Spawns
+  suffix: Pests Only
   components:
    - type: VentCritterSpawnLocation
      canSpawn: false
